Me typing thoughts
I don't care if we keep or scrap this. just putting down thoughts.

My attempt is to make this something of a story for a user to follow. I
think dumbing the language down a bit will make for a more pleasing
video.

Feel free to contradict me at any point in time. I am NOT a marketing
person and I know my limitations.